requirements.txt
python-dotenv==1.0.0
Requests==2.31.0
webuntis==0.1.21
